+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSMineshaftEndStructure.java
@@ -17,13 +17,9 @@ import net.minecraft.world.gen.feature.structure.Structure;
 import java.util.List;
 
 
-public class RSMineshaftEndStructure extends AbstractMineshaftStructure {
-    public RSMineshaftEndStructure(Codec<NoFeatureConfig> config) {
-        super(config);
-    }
-
-    public double getProbability() {
-        return RepurposedStructures.RSMineshaftsConfig.endMineshaftSpawnrate.get();
+public class RSMineshaftEndStructure extends RSMineshaftStructure {
+    public RSMineshaftEndStructure(Codec<NoFeatureConfig> config, RSMineshaftPieces.Type mineshaftType, double probability, int maxHeight, int minHeight) {
+        super(config, mineshaftType, probability, maxHeight, minHeight);
     }
 
     private static final List<Biome.SpawnListEntry> MONSTER_SPAWNS = Lists.newArrayList(
@@ -36,35 +32,10 @@ public class RSMineshaftEndStructure extends AbstractMineshaftStructure {
     }
 
     @Override
-    public Structure.IStartFactory<NoFeatureConfig> getStartFactory() {
-        return RSMineshaftEndStructure.Start::new;
-    }
-
-    @Override
     protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeProvider biomeSource, long seed, SharedSeedRandom chunkRandom, int x, int z, Biome biome, ChunkPos chunkPos, NoFeatureConfig featureConfig) {
         chunkRandom.setLargeFeatureSeed(seed, x, z);
-        double d = (getProbability() / 10000D);
+        double d = (probability / 10000D);
         int landHeight = chunkGenerator.func_222531_c(x, z, Heightmap.Type.WORLD_SURFACE_WG);
         return chunkRandom.nextDouble() < d && (RepurposedStructures.RSMineshaftsConfig.barrensIslandsEndMineshafts.get() || landHeight > 20);
     }
-
-    public static class Start extends AbstractStart {
-        public Start(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
-            super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
-        }
-
-        @Override
-        public RSMineshaftPieces.Type getMineshaftType() {
-            return RSMineshaftPieces.Type.END;
-        }
-        @Override
-        public int getMaxHeight() {
-            return RepurposedStructures.RSMineshaftsConfig.endMineshaftMaxHeight.get();
-        }
-        @Override
-        public int getMinHeight() {
-            return RepurposedStructures.RSMineshaftsConfig.endMineshaftMinHeight.get();
-        }
-    }
-
 }