+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/RSMineshaftStructure.java
@@ -14,40 +14,47 @@ import net.minecraft.world.gen.feature.structure.StructureStart;
 import net.minecraft.world.gen.feature.template.TemplateManager;
 
 
-public abstract class AbstractMineshaftStructure extends AbstractBaseStructure {
+public class RSMineshaftStructure extends AbstractBaseStructure {
+    protected final double probability;
+    protected final int maxHeight;
+    protected final int minHeight;
+    protected final RSMineshaftPieces.Type mineshaftType;
 
-    public AbstractMineshaftStructure(Codec<NoFeatureConfig> config) {
+    public RSMineshaftStructure(Codec<NoFeatureConfig> config, RSMineshaftPieces.Type mineshaftType, double probability, int maxHeight, int minHeight) {
         super(config);
+        this.probability = probability;
+        this.maxHeight = maxHeight;
+        this.minHeight = minHeight;
+        this.mineshaftType = mineshaftType;
     }
-
-    public abstract double getProbability();
 
     @Override
     protected boolean shouldStartAt(ChunkGenerator chunkGenerator, BiomeProvider biomeSource, long seed, SharedSeedRandom chunkRandom, int x, int z, Biome biome, ChunkPos chunkPos, NoFeatureConfig featureConfig) {
         chunkRandom.setLargeFeatureSeed(seed, x, z);
-        double d = (getProbability() / 10000D);
+        double d = (probability / 10000D);
         return chunkRandom.nextDouble() < d;
     }
 
-    public static abstract class AbstractStart extends StructureStart<NoFeatureConfig> {
-        public AbstractStart(Structure<NoFeatureConfig> feature, int chunkX, int chunkZ, MutableBoundingBox box, int references, long seed) {
+    @Override
+    public IStartFactory<NoFeatureConfig> getStartFactory() {
+        return Start::new;
+    }
+
+    public class Start extends StructureStart<NoFeatureConfig> {
+        public Start(Structure<NoFeatureConfig> feature, int chunkX, int chunkZ, MutableBoundingBox box, int references, long seed) {
             super(feature, chunkX, chunkZ, box, references, seed);
         }
 
-        public abstract RSMineshaftPieces.Type getMineshaftType();
-        public abstract int getMaxHeight();
-        public abstract int getMinHeight();
-
         @Override
         public void init(ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig NoFeatureConfig) {
-            RSMineshaftPieces.Room structuremineshaftpiecesua$room = new RSMineshaftPieces.Room(0, this.rand, (chunkX << 4) + 2, (chunkZ << 4) + 2, getMineshaftType());
+            RSMineshaftPieces.Room structuremineshaftpiecesua$room = new RSMineshaftPieces.Room(0, this.rand, (chunkX << 4) + 2, (chunkZ << 4) + 2, mineshaftType);
             this.components.add(structuremineshaftpiecesua$room);
 
             structuremineshaftpiecesua$room.buildComponent(structuremineshaftpiecesua$room, this.components, this.rand);
             this.recalculateStructureSize();
 
-            int minimum = getMinHeight();
-            int maximum = Math.max(getMaxHeight(), minimum) + 1;
+            int minimum = minHeight;
+            int maximum = Math.max(maxHeight, minimum) + 1;
 
             int offset = this.rand.nextInt(maximum - minimum) + minimum;
             this.bounds.offset(0, offset - 50, 0);