+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/VillageSwampStructure.java
@@ -2,51 +2,33 @@ package com.telepathicgrunt.repurposedstructures.world.structures;
 
 import com.mojang.serialization.Codec;
 import com.telepathicgrunt.repurposedstructures.RSFeatures;
-import com.telepathicgrunt.repurposedstructures.RepurposedStructures;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.SharedSeedRandom;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.MutableBoundingBox;
+import net.minecraft.world.IWorldReader;
 import net.minecraft.world.biome.Biome;
+import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
-import net.minecraft.world.gen.feature.structure.Structure;
+import net.minecraft.world.gen.feature.structure.*;
 import net.minecraft.world.gen.feature.template.TemplateManager;
+import net.minecraft.world.gen.settings.StructureSeparationSettings;
 
-public class VillageSwampStructure extends AbstractVillageStructure
-{
-    public VillageSwampStructure(Codec<NoFeatureConfig> config) {
-	super(config);
+public class VillageSwampStructure extends VillageBaseStructure {
+    public VillageSwampStructure(Codec<NoFeatureConfig> config, ResourceLocation poolRL, int structureSize) {
+        super(config, poolRL, structureSize);
     }
 
-	@Override
-	public Structure<NoFeatureConfig> getVillageInstance() {
-		return RSFeatures.SWAMP_VILLAGE;
-	}
+    public class Start extends MainStart {
+        public Start(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
+            super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
+        }
 
-    public Structure.IStartFactory<NoFeatureConfig> getStartFactory() {
-		return VillageSwampStructure.Start::new;
-    }
-
-    public static class Start extends AbstractStart
-    {
-		public Start(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
-			super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
-		}
-
-		public static ResourceLocation VILLAGE_ResourceLocation = new ResourceLocation(RepurposedStructures.MODID + ":village/swamp/town_centers");
-		@Override
-		public ResourceLocation getResourceLocation() {
-			return VILLAGE_ResourceLocation;
-		}
-
-		@Override
-		public int getSize() {
-			return 6;
-		}
-
-		@Override
-		public void init(ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig NoFeatureConfig) {
-			super.init(chunkGenerator, structureManager, chunkX, chunkZ, biome, NoFeatureConfig);
-			this.components.get(0).offset(0,-1,0);
-		}
+        public void init(ChunkGenerator chunkGenerator, TemplateManager structureManager, int chunkX, int chunkZ, Biome biome, NoFeatureConfig NoFeatureConfig) {
+            super.init( chunkGenerator, structureManager, chunkX, chunkZ, biome, NoFeatureConfig);
+            this.components.get(0).offset(0, -1, 0);
+        }
     }
 }